pyproject.toml
[project]
name = "uv-work"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles==24.1.0",
    "aiohappyeyeballs==2.4.4",
    "aiohttp==3.11.11",
    "aioshutil==1.5",
    "aiosignal==1.3.2",
    "annotated-types==0.7.0",
    "anyio==4.7.0",
    "appnope==0.1.4",
    "arxiv==2.1.3",
    "asgiref==3.8.1",
    "asttokens==3.0.0",
    "attrs==24.3.0",
    "backoff==2.2.1",
    "bcrypt==4.2.1",
    "beautifulsoup4==4.12.3",
    "boto3==1.35.90",
    "botocore==1.35.90",
    "build==1.2.2.post1",
    "cachetools==5.5.0",
    "certifi==2024.12.14",
    "cffi==1.17.1",
    "chardet==5.2.0",
    "charset-normalizer==3.4.1",
    "chroma-hnswlib==0.7.6",
    "chromadb==0.5.23",
    "click==8.1.8",
    "coloredlogs==15.0.1",
    "comm==0.2.2",
    "cryptography==44.0.0",
    "dataclasses-json==0.6.7",
    "debugpy==1.8.11",
    "decorator==5.1.1",
    "deprecated==1.2.15",
    "distro==1.9.0",
    "duckduckgo-search==7.1.1",
    "durationpy==0.9",
    "emoji==2.14.0",
    "eval-type-backport==0.2.2",
    "executing==2.1.0",
    "fastapi==0.115.6",
    "feedparser==6.0.11",
    "filelock==3.16.1",
    "filetype==1.2.0",
    "flatbuffers==24.12.23",
    "frozendict==2.4.6",
    "frozenlist==1.5.0",
    "fsspec==2024.12.0",
    "google-api-core==2.24.0",
    "google-api-python-client==2.156.0",
    "google-auth==2.37.0",
    "google-auth-httplib2==0.2.0",
    "google-auth-oauthlib==1.2.1",
    "google-cloud-core==2.4.1",
    "googleapis-common-protos==1.66.0",
    "grpcio==1.68.1",
    "h11==0.14.0",
    "html2text==2024.2.26",
    "html5lib==1.1",
    "httpcore==1.0.7",
    "httplib2==0.22.0",
    "httptools==0.6.4",
    "httpx==0.27.2",
    "httpx-sse==0.4.0",
    "huggingface-hub==0.27.0",
    "humanfriendly==10.0",
    "idna==3.10",
    "importlib-metadata==8.5.0",
    "importlib-resources==6.4.5",
    "ipykernel==6.29.5",
    "ipython==8.31.0",
    "jedi==0.19.2",
    "jinja2==3.1.5",
    "jiter==0.8.2",
    "jmespath==1.0.1",
    "joblib==1.4.2",
    "jsonpatch==1.33",
    "jsonpath-python==1.0.6",
    "jsonpointer==3.0.0",
    "jsonschema==4.23.0",
    "jsonschema-specifications==2024.10.1",
    "jupyter-client==8.6.3",
    "jupyter-core==5.7.2",
    "kubernetes==31.0.0",
    "langchain==0.3.13",
    "langchain-chroma==0.1.4",
    "langchain-community==0.3.13",
    "langchain-core>=0.3.56",
    "langchain-experimental==0.3.4",
    "langchain-google-community==2.0.3",
    "langchain-mcp-adapters>=0.1.0",
    "langchain-ollama>=0.3.2",
    "langchain-openai==0.2.14",
    "langchain-text-splitters==0.3.4",
    "langdetect==1.0.9",
    "langgraph==0.2.60",
    "langgraph-checkpoint==2.0.9",
    "langgraph-sdk==0.1.48",
    "langsmith==0.2.6",
    "litellm==1.56.3",
    "lxml==5.3.0",
    "markdown==3.7",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "marshmallow==3.23.2",
    "matplotlib-inline==0.1.7",
    "mcp[cli]>=1.6.0",
    "mdurl==0.1.2",
    "mmh3==5.0.1",
    "monotonic==1.6",
    "mpmath==1.3.0",
    "msgpack==1.1.0",
    "multidict==6.1.0",
    "multitasking==0.0.11",
    "mypy-extensions==1.0.0",
    "nest-asyncio==1.6.0",
    "nltk==3.9.1",
    "numpy==1.26.4",
    "oauthlib==3.2.2",
    "olefile==0.47",
    "ollama>=0.5.1",
    "onnxruntime==1.20.1",
    "openai==1.58.1",
    "opentelemetry-api==1.29.0",
    "opentelemetry-exporter-otlp-proto-common==1.29.0",
    "opentelemetry-exporter-otlp-proto-grpc==1.29.0",
    "opentelemetry-instrumentation==0.50b0",
    "opentelemetry-instrumentation-asgi==0.50b0",
    "opentelemetry-instrumentation-fastapi==0.50b0",
    "opentelemetry-proto==1.29.0",
    "opentelemetry-sdk==1.29.0",
    "opentelemetry-semantic-conventions==0.50b0",
    "opentelemetry-util-http==0.50b0",
    "orjson==3.10.12",
    "overrides==7.7.0",
    "packaging==24.2",
    "pandas==2.2.3",
    "parso==0.8.4",
    "pdf2image==1.17.0",
    "peewee==3.17.8",
    "pexpect==4.9.0",
    "pillow==11.0.0",
    "platformdirs==4.3.6",
    "posthog==3.7.4",
    "primp==0.9.2",
    "prompt-toolkit==3.0.48",
    "propcache==0.2.1",
    "proto-plus==1.25.0",
    "protobuf==5.29.2",
    "psutil==6.1.1",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.3",
    "py-zerox==0.0.7",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.1",
    "pycparser==2.22",
    "pydantic==2.9.2",
    "pydantic-core==2.23.4",
    "pydantic-settings==2.7.0",
    "pygments==2.18.0",
    "pyparsing==3.2.0",
    "pypdf==5.1.0",
    "pypdf2==3.0.1",
    "pypika==0.48.9",
    "pyproject-hooks==1.2.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "python-iso639==2024.10.22",
    "python-magic==0.4.27",
    "python-oxmsg==0.0.1",
    "pytz==2024.2",
    "pyyaml==6.0.2",
    "pyzmq==26.2.0",
    "rapidfuzz==3.11.0",
    "referencing==0.35.1",
    "regex==2024.11.6",
    "requests==2.32.3",
    "requests-oauthlib==2.0.0",
    "requests-toolbelt==1.0.0",
    "rich==13.9.4",
    "rpds-py==0.22.3",
    "rsa==4.9",
    "s3transfer==0.10.4",
    "sgmllib3k==1.0.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "soupsieve==2.6",
    "sqlalchemy==2.0.36",
    "stack-data==0.6.3",
    "starlette==0.41.3",
    "streamlit>=1.43.0",
    "sympy==1.13.3",
    "tenacity==9.0.0",
    "tiktoken==0.8.0",
    "tokenizers==0.20.3",
    "tornado==6.4.2",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "typer==0.15.1",
    "typing-extensions==4.12.2",
    "typing-inspect==0.9.0",
    "tzdata==2024.2",
    "unstructured==0.16.11",
    "unstructured-client==0.28.1",
    "uritemplate==4.1.1",
    "urllib3==2.3.0",
    "uvicorn==0.34.0",   
    "watchfiles==1.0.3",
    "wcwidth==0.2.13",
    "webencodings==0.5.1",
    "websocket-client==1.8.0",
    "websockets==14.1",
    "wrapt==1.17.0",
    "yarl==1.18.3",
    "yfinance==0.2.51",
    "zipp==3.21.0",
    "docx2txt==0.8",
]
